Move cell reference tests to separate file
